+++ CHANGELOG.md
@@ -24,6 +24,7 @@ BUG FIXES:
 - Add `data_salt` and `signature_sign_method` properties to data from AS when querying on RP side.
 - Fix error in `getMessageWithCode()` in CustomError when error cause is undefined.
 - Clean up data in cache DB when create request and create identity fails.
+- Fix AS can send data response with any request ID and service ID without having to receive the request first.
 
 ## 0.5.4 (July 27, 2018)